Add Graphql Client (#11)
* Add Apollo Client

Add the client and a fledgling query.  This is a WIP,
and queries will become cleaner with iteration.

Still need to wire up schema download and code gen
with local Gradle builds as well as CI.

* Generate Schema in CI

Generate the grapql schema from GitHub at build
time.

* Some CI Cleanup

Remove the usage of the apollo branch in CI for testing.

Also, add a badge for ktlint to represent!

* Add Documentation for Build

Add a little documentation on how to build
from the source now that there is a dependency
on the apollo client and the github schema.